% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\docType{data}
\name{platforms.accession}
\alias{platforms.accession}
\title{Platform Data}
\format{A character object with:
\describe{
  \item{platform.accession}{GSE Numbers}

}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/browse/?view=platforms}
}
\usage{
platforms.accession
}
\description{
A dataset containing the platform meta data needed for
populating the search box.
}
\keyword{datasets}
