% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\docType{data}
\name{platforms.description}
\alias{platforms.description}
\title{Platform Data}
\format{A character object with:
\describe{
  \item{platform.description}{Description Title}

}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/browse/?view=platforms}
}
\usage{
platforms.description
}
\description{
A dataset containing the platform meta data needed for
populating the search box.
}
\keyword{datasets}
