% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshinyGEO.R
\name{runshinyGEO}
\alias{runshinyGEO}
\title{Run shinyGEO application}
\usage{
runshinyGEO()
}
\value{
Nothing is returned except a shiny app launched locally showing
shinyGEO.
}
\description{
Run shinyGEO application
}
\examples{
\dontrun{ runshinyGEO() }
}
