% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\docType{data}
\name{series.description}
\alias{series.description}
\title{Series Data}
\format{A character object with:
\describe{
  \item{series.description}{Description Title}

}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/browse/}
}
\usage{
series.description
}
\description{
A dataset containing the Series meta data needed for
populating the search box.
}
\keyword{datasets}
