% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGEO.R
\docType{package}
\name{shinyGEO}
\alias{shinyGEO}
\alias{shinyGEO-package}
\alias{shinyGEO-package}
\title{shinyGEO: A package for analyzing gene expression expression data with shiny.}
\description{
The shinyGEO package provides two important functions:
updateGEOdata, runshinyGEO.
}
\section{shinyGEO functions}{

updateGEOdata() updates the meta-data for populating the search
box for series and platform data from the Gene Expression Omnibus.
runshinyGEO() launches the shiny application locally.
}

