% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_GEOdata.R
\name{update_GEOdata}
\alias{update_GEOdata}
\title{Update series and platform data from GEO}
\usage{
update_GEOdata()
}
\value{
Series and platform data are updated from the Gene Expression
Omnibus and stored as R data objects.
}
\description{
Update series and platform data from GEO
}
\examples{
\dontrun{ update_GEOdata() }
}
