% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGEOdata.R
\name{updateGEOdata}
\alias{updateGEOdata}
\title{Update series and platform data from GEO}
\usage{
updateGEOdata()
}
\value{
Series and platform metadata are updated from the Gene Expression
Omnibus and stored as R data objects to create a 'database' of
valid accession numbers.
}
\description{
Update series and platform data from GEO
}
\examples{
\dontrun{ updateGEOdata() }
}
